Refactor token generation and add login feature test
Introduced GerarTokenController for token generation, updated GerarTokenResquest model fields and package, removed the old GerarToken class, and added a new login.feature file for validating login functionality.
